+++ lib-hal/include/rpi/console_fb.h
@@ -49,7 +49,7 @@ extern void console_puthex_fg_bg(uint8_t, uint16_t, uint16_t);
 extern void console_putpct_fg_bg(uint8_t, uint16_t, uint16_t);
 extern void console_put3dec_fg_bg(uint8_t, uint16_t, uint16_t);
 
-extern int console_status(uint16_t, const char *);
+extern int console_status(uint32_t, const char *);
 
 #ifdef __cplusplus
 }
